Refactor message sending functions and remove deprecated database connection code
